refactor: extract python version determining funcs
